[FLINK-5129] do not check for file existence before copying from it
There are file systems where this imposes a significant cost compared to just
trying to read from a file. Also if the file did not exist, the underlying
fs implementation should throw a meaningful exception itself.

+++ flink-runtime/src/main/java/org/apache/flink/runtime/blob/FileSystemBlobStore.java
@@ -102,25 +102,20 @@ public class FileSystemBlobStore implements BlobStore {
 
 		final Path fromPath = new Path(fromBlobPath);
 
-		try {
-			if (fileSystem.exists(fromPath)) {
-				try (InputStream is = fileSystem.open(fromPath);
-					FileOutputStream fos = new FileOutputStream(toFile)) {
-					LOG.debug("Copying from {} to {}.", fromBlobPath, toFile);
-					IOUtils.copyBytes(is, fos); // closes the streams
-				}
-			} else {
-				throw new IOException(fromBlobPath + " does not exist.");
-			}
-		} catch (Exception e) {
+		boolean success = false;
+		try (InputStream is = fileSystem.open(fromPath);
+			FileOutputStream fos = new FileOutputStream(toFile)) {
+			LOG.debug("Copying from {} to {}.", fromBlobPath, toFile);
+			IOUtils.copyBytes(is, fos); // closes the streams
+			success = true;
+		} finally {
 			// if the copy fails, we need to remove the target file because
 			// outside code relies on a correct file as long as it exists
-			try {
-				toFile.delete();
-			} catch (Exception e1) {
-				// ignore
+			if (!success) {
+				try {
+					toFile.delete();
+				} catch (Throwable ignored) {}
 			}
-			throw e;
 		}
 	}
